Fix publish of checksum files
When the checksum files are already exist in the bin directory,
duplicate paths were recorded, and included in the gh release.

This caused creating a release to fail.

I have updated the logic to only get a list of unique files. I've also
changed the release creation logic so that we aways create the release
and upload the assets separately. The error handling in the upload
command seems to be better than in the release create command.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>