Only download porter runtime when porter changes
I was checking if the porter-runtime version matched the desired version and completely forgot that check will _only_ pass on linux. So on most local dev machines, win/mac, it was always downloading the runtime each build because the version check failed. You can't check the version when the architecture is a mismatch...

I've updated the EnsurePorter logic so that we download the runtime binary when:
* It's missing
* We downloaded the client

This way repeated builds won't re-download anything, but if the change the desired client version that will force the runtime to download too.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>