Calculate version correctly when multiple tags are present
When multiple tags are present on the same commit, the version was
calculated incorrectly. E.g., when `latest`, `canary` and `v1.1.0` tags
are present on the same commit, the command seems to sort alphabetically
causing the version to always be set to `canary`. This behaviour breaks
makes it impossible to release a finalized version, e.g., `v1.1.0`.

Using matching results in the version being calculated from the latest
released version. The version will look like this `v1.1.0-2-gf3862833`
instead of `canary`.

Signed-off-by: Kim Christensen <kmch@simcorp.com>